Message.Meta is deprecated in favour of Message.GetMeta() and friends #1.
This commit introduces a family of Message.meta-related methods.
The motivation for the change is describe in the issue #1, where msg
meta experienced a concurrent modification and access and caused a
daemin crash. This change hides access to message Meta and replaces the
internals of meta with sync.Map. On top if it, the methods use a mutex
which is supposed to protect agains a concurrent access.

Methods list:
  * GetMeta(key string) (interface{}, bool)
  * GetMetaOrDef(key string, def interface{}) (interface{}, bool)
  * GetMetaAll() map[string]interface{}
  * SetMeta(key string, val interface{})
  * SetMetaAll(map[string]interface{})
  * UnsetMeta(key string) (interface{}, bool)
  * UnsetMetaAll() map[string]interface{}