Massive constructor function refactoring.
All core link constructor functions are called New(), with no
exceptions. This is done in order to unify the naming conventions and
make all constructors look the same. This does not directly imply on
plugins constructor function naming. If a plugin is supposed to
instantiate multiple distinct links, it still make sense to call them
something like NewPluginLink1, NewPluginLink2 etc.