Reworked global config registry so it uses a fallback chain #4
This commit changes the way config values are resolved.
The first solution was based on a wrong assumption that if a provider
registers itself under a specific config key, it will definitely serve
it. A good counter example of it is CLI-ENV-default providers chain. CLI
and ENV register themselves and there is no guarantee a user will
provide these settings. In this case if propagates down to default
provider.

In fact, this change turns a config lookup operation to O(N) due to the
propagation process. This doesn't seem to be a big deal as the global
registry is there to resolve multiple startup-time config collisions and
is mostly needed to tear up the processes, not to serve millions of
lookups per second. Read: it's slow by design, and trading O(1) to O(N)
seems to be a decent solution.