Fixed corrupted data problem in UDP sender #5
The problem was caused by 2 major reasons.
First, the easy one, net_sender was sending empty messages and this
confused the receiver. Evio has a non-zero message length filter, so it
worked fine.

Second, the tough one: bufio reader has 2 very different methods:
ReadLine and ReadBytes. The first one returns a slice referencing the current
buffer, whereas the second returns a copy of this buffer. Another
problem was that I ws not taking care of the buffer completeness:
sometimes a buffer can not be assembled form a single go. Changing the
way the data is read form the buffer solved the problem, even though it
made the pipeline roughly as twice as slow. More performnce tuning to
come.