Fixed sending acks to a closed channel in replicator link.
This commit changes the way we await for ack notifications from message
copies in replicator link. In the past, it was a fire-and-forget
approach: we used defer with no real awiating. We keep using WaitGroups,
but we resolve them when a notification from msgCp was received. This
gives us a real idea about message status and the moment of it's
resolution.