Implemented mpx.Multiplex: generic multiplexing function.
This commit introduces a generic Multiplex function which accepts a
message, a collection of core.Link and a timeout and performs a decent
multiplexing with a composite status report. This is planned to be used
by mpx itself and by a set of plugins where a problem of multiplexed
submission comes up.