Renamed mpx and dmx to demux and mux accordingly
Massive refactoring, finally fixed the naming: mux was always meant to
be called demultiplexer, and dmx was supposed to be renamed to
multiplexer. This has not taken place. Some significant speedups in
demux.Demultiplex (former mpx.Multiplex): if a message is not marked as
synchronous, it will never be waited to be finalized (speedup up to
100%). Another optimization in the same method: instead of constant
copying the links around, it builds a select-mask and picks the relevant
nodes. This comes with a limitation: up to 64 connected links supported
only (uint64). This provided a speedup up to 10%.