Replicator and Demuxer use threadiness setting to parallelize
This commit changes the default behavior of replicator and demuxer: they
read max threadiness setting and spawn as many parallel channer readers
and multiplexers and many threads are enabled by the system config. This
gives a significant (up to 25%) speed improvement to the system
performance. It's been noticed the top performance of the system is
reached with GOMAXPROCS=2 and it slowly degrades with the growing amount
of threads.