Dev env fragments + flow-tool docker compose command (#33)
* Completed the draft version of flow-tool docker compose command
* Created the skeleton for flow-tool
* Implemented devenv package which will be responsible for dev env setup.

`devenv` package comes as a set of classes providing a dev env setup
toolbox. Links should implement a public method called DevEnv() which
takes an instance of devenv.Context and returns a list of devenv
Fragments. A fragment is an abstract piece of puzzle which has some
specific type (DockerComposeFragment is the only one existing at the
moment). Links use Context in order to pass the data around. Say, if
there are multiple siblings, they can pass around a port number used by
the last instance, or the last generated name in order to produce a
sequentually following items.
* Implemented udp and tcp sink docker-compose fragments and extended Fragments with DockerComposeBuilder func
* Implemented a simple config provider, moved core test utils to a separate utils/core_test package and did a refactoring